[py3] Cast all division results back to int
When connecting to Live 11.2.6 with my Launchpad Pro mk1, the sequencer throws
errors related to unexpected float types for several of the side buttons (eg.
change octave, enter melodic mode).

This commit attempts to find all places where division is used and a floating
point number is returned instead of an int (Python 3 changes the default
behavior of int division).

I've tested the step sequencer mode and all features seem to work. I'm unable to
reproduce any of the errors in the logs now.